Refactor tests to use description instead of test index
When using test case index to identify specific (possibly failing) test,
the implementation can be a bit unambiguous and occasionally hard to
track. There are questions whether the index counting starts from 0 or 1
and whether it's present in comment (and keeping comments up-to-date) or
not etc.

Using description field and sub-tests make it clear how test cases are
identified. It also gives specific test cases meaningful name when
running them.