Rename current Cluster types to LegacyCluster
As a preparation for introducing controller to reconcile Cluster CR, rename
current Cluster types to LegacyCluster.

Plan here is to eventually reconcile only on Cluster CR and have configuration
to extract common part of provider specific extension of it. This way we can
simplify the code and make it a bit more straightforward.

In the first phase the Cluster CR reconciliation is used to map information
from Cluster CR to AWSClusterConfig and drive cluster-operator logic that way.