Add initial awsclusterconfig resources execution conditionals
In order to have single controller for different providers, it makes sense to
have a separate resource implementation for each provider specific
clusterconfig mapping and in the beginning of Ensure{Created,Deleted}
functions, check if the given Cluster CR contains the right provider extension.

This way the implementation is pretty clean and can be run on all providers
without any special configuration etc.