Add mapping of workers from Cluster -> AWSClusterConfig (#523)
Add basic mapping of workers (from MachineDeployments) to AWSClusterConfig so
that resources depending on e.g. node count and such information work as
before. The structure in AWSClusterConfig follows the old one, but should do
for us for now. We should get rid of that once we drop the cluster config CRs
eventually.

Also added some required fixes for machinedeploymentstatus resource.