Do not update app-operator.giantswarm.io/version label of unique app-operator managed App CRs
The label updating mechanism is there to handle cluster upgrades when the cluster components
get updated, e.g. app-operator was updated from 5.8.0 to 5.9.0. In that case the normal WC
App CRs needs to have their label updated so the new 5.9.0 app-operator will start handling
them.

However app-of-apps should be managed by the unique app-operator that has the label 0.0.0
because WC app-opearator are lacking permissions to do so. So in this case the - or when
otherwise explicitly set for an App CR to be manged by the unique operator - it should
not force the App CR to be managed by the WC app-operator.